feat: split NetworkManager and TransformRequestManager
