+++ package/ios/modules/transform-request/MLRNTransformRequestModule.h
@@ -1,0 +1,9 @@
+#import <MapLibreReactNativeSpec/MapLibreReactNativeSpec.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface MLRNTransformRequestModule : NativeTransformRequestModuleSpecBase <NativeTransformRequestModuleSpec>
+@end
+
+NS_ASSUME_NONNULL_END
+
