+++ package/src/modules/network/NativeNetworkModule.ts
@@ -1,0 +1,8 @@
+import type { TurboModule } from "react-native";
+import { TurboModuleRegistry } from "react-native";
+
+export interface Spec extends TurboModule {
+  setConnected(connected: boolean): void;
+}
+
+export default TurboModuleRegistry.getEnforcing<Spec>("MLRNNetworkModule");
